Introduce JsonIndexAccessor
This commit introduces a JsonIndexAccessor as a complement to the
existing JsonPropertyAccessor.

When a JsonIndexAccessor is registered with the SpEL EvaluationContext,
JSON arrays can be consistently indexed via integer literals (e.g.,
[1]) instead of string literals representing integers (e.g., ['1']).

See gh-9383
